+++ src/scicloj/kindly/v1/kind.clj
@@ -1,0 +1,1 @@
+(ns scicloj.kindly.v1.kind)