+++ src/scicloj/kindly/v2/kind.clj
@@ -1,0 +1,1 @@
+(ns scicloj.kindly.v2.kind)