+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [2-alpha1] - 2022-03-28
+- various API changes, in particular the Kindness protocol
+
 ## [1-alpha3] - 2021-11-12
 - added a `to-hiccup` function to the api